requirements.txt
flask
pillow
pytesseract
indic-transliteration
